+++ packages/agentmark/src/adapters/default.ts
@@ -10,41 +10,29 @@ import {
   AdapterImageResult
 } from "../types";
 import type { Schema } from 'ai';
+import { jsonSchema } from 'ai';
 
 // Define specific return types for DefaultAdapter
 export type DefaultTextResult<T = string> = TextConfig & AdapterTextResult<T>;
-export type DefaultObjectResult<T = unknown> = ObjectConfig & AdapterObjectResult<T>;
+export type DefaultObjectResult<T = unknown> = ObjectConfig & AdapterObjectResult<T> & { schema?: Schema<T> };
 export type DefaultImageResult<T = string> = ImageConfig & AdapterImageResult<T>;
 
 export class DefaultAdapter implements Adapter<DefaultTextResult<any>, DefaultObjectResult<any>, DefaultImageResult<any>> {
   adaptText<T = string>(
     input: TextConfig, 
-    options?: JSONObject,
-    settings?: PromptMetadata
   ): DefaultTextResult<T> {
-    return {
-      ...input
-    };
+    return input;
   }
 
   adaptObject<T = unknown>(
     input: ObjectConfig & { jsonSchema?: Schema<T> },
-    options?: JSONObject,
-    settings?: PromptMetadata
   ): DefaultObjectResult<T> {
-    return { 
-      ...input, 
-      schema: input.jsonSchema 
-    };
+    return input;
   }
 
   adaptImage<T = string>(
     input: ImageConfig,
-    options?: JSONObject,
-    settings?: PromptMetadata
   ): DefaultImageResult<T> {
-    return {
-      ...input
-    };
+    return input;
   }
 }
