+++ packages/agentmark/src/agentmark.ts
@@ -1,6 +1,8 @@
 import { Loader, TemplateEngine, Adapter } from "./types";
-import { TextPrompt, ImagePrompt, Prompt, TextPromptInterface, ImagePromptInterface } from "./prompts";
+import { Prompt } from "./prompts";
+import { TextPrompt, TextPromptInterface } from "./prompts/text";
 import { ObjectPrompt, ObjectPromptInterface } from "./prompts/object";
+import { ImagePrompt, ImagePromptInterface } from "./prompts/image";
 import { TextConfigSchema, ObjectConfigSchema, ImageConfigSchema } from "./schemas";
 import { TemplatedxTemplateEngine } from "./template_engines/templatedx";
 
@@ -82,39 +84,5 @@ export class AgentMark<
     
     ImageConfigSchema.parse(await this.templateEngine.compile(content));
     return new ImagePrompt<T[K]["input"], T[K]["output"], A>(content, this.templateEngine, this.adapter, pathOrPreloaded);
-  }
-
-  /**
-   * Generic method to load any type of prompt
-   */
-  async loadPrompt<K extends keyof T & string>(
-    type: 'text',
-    pathOrPreloaded: K
-  ): Promise<TextPromptInterface<T[K]["input"], T[K]["output"], A>>;
-  
-  async loadPrompt<K extends keyof T & string>(
-    type: 'object',
-    pathOrPreloaded: K
-  ): Promise<ObjectPromptInterface<T[K]["input"], T[K]["output"], A>>;
-  
-  async loadPrompt<K extends keyof T & string>(
-    type: 'image',
-    pathOrPreloaded: K
-  ): Promise<ImagePromptInterface<T[K]["input"], T[K]["output"], A>>;
-  
-  async loadPrompt<K extends keyof T & string>(
-    type: 'text' | 'object' | 'image',
-    pathOrPreloaded: K
-  ): Promise<Prompt<T[K]["input"], T[K]["output"], A>> {
-    switch (type) {
-      case 'text':
-        return this.loadTextPrompt(pathOrPreloaded);
-      case 'object':
-        return this.loadObjectPrompt(pathOrPreloaded);
-      case 'image':
-        return this.loadImagePrompt(pathOrPreloaded);
-      default:
-        throw new Error(`Unknown prompt type: ${type}`);
-    }
   }
 }
